+++ lib/acts_as_favoritor/configuration.rb
@@ -1,34 +1,13 @@
 module ActsAsFavoritor
+    class Configuration
 
-    def self.default_scope
-        config = get_config
-        if config&.key :default_scope
-            config[:default_scope]
-        else
-            'favorite'
+        attr_accessor :default_scope
+        attr_accessor :cache
+
+        def initialize
+            @default_scope = 'favorite'
+            @cache = false
         end
+
     end
-
-    def self.cache
-        config = get_config
-        if config&.key :cache
-            config[:cache]
-        else
-            false
-        end
-    end
-
-
-    private
-
-
-    def self.get_config
-        require 'yaml'
-
-        begin
-            YAML.load_file 'config/acts_as_favoritor.yml'
-        rescue Exception
-        end
-    end
-
 end